cleanup: no more actor.s
